backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0

pydantic==2.10.4
pydantic-settings==2.7.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

httpx==0.28.1
authlib==1.3.2

redis==5.2.1

boto3==1.35.86

